Change the way Codable macro works
Codable macro will now suggests a fix-it which replace Codable with Decodable and Encodable macros.

The reason for this is to avoid maintaining Codable macro code because it is essentially a custom logic of Decodable and Encodable macro.

In the future when macro supports typealias or allows peer macro to modify the attributes list of a type declaration, Codable macro can perform the replacement behind the scene without the fix-it suggestion.